i386_64 fixes
git-svn-id: file:///media/bill/diskAshur2/backup/svn/microemacs/releases/me_20060909@1871 a259ec66-be10-11ea-b557-b34f8aa0b108

+++ src/registry.c
@@ -804,7 +804,7 @@ regDecodeMode (int *modep, meUByte *modeStr)
         {
             if((ss=meStrchr(meRegModeList,toLower(cc))) == NULL)
                 return mlwrite(MWABORT|MWPAUSE,(meUByte *)"Illegal Registry mode [%c]",cc) ;
-            jj = (1<< (((int) ss)-((int) meRegModeList))) ;
+            jj = (1 << ((int)(ss - meRegModeList))) ;
             if(isUpper(cc))
                 mode &= ~jj ;
             else
